Move Transport Graph JSON-serialization test
